Kick from team test implemented and passes and refactored setup and tear down team test code into separate file
