+++ app/api/src/tests/teams.test.ts
@@ -1,131 +1,53 @@
-import { joinTeam, createTeam  } from '../features/teams/teams.controller';
+import { joinTeam, kickTeamMember  } from '../features/teams/teams.controller';
 import prisma from '../config/prismaClient';
 import type { Event, Team, TeamMember } from '@prisma/client';
 import { describe, beforeEach, afterEach, expect, test } from 'vitest';
+import { cleanTeamsTestUp, setTeamsTestUp } from './teamsTestSetupTearDown';
 
-interface TeamInfo {
+describe('Teams Tests', ()=>{
+  
+  describe('Team tests where there are 2 teams with one member on each team', ()=> {
+    let teamSetupInfo: TeamTestSetupInfo;
+    beforeEach(async () => {
+      teamSetupInfo = await setTeamsTestUp([1, 1]);
+    });
+
+    afterEach(async () => {
+      await cleanTeamsTestUp(teamSetupInfo);
+    });
+
+    test('join team test', async ()=>{
+      await testJoinTeamOneMemPerTeam(teamSetupInfo);
+    });
+  });
+
+  describe('Team tests where there is only one team with multiple members', ()=> {
+    let teamSetupInfo: TeamTestSetupInfo;
+    beforeEach(async () => {
+      teamSetupInfo = await setTeamsTestUp([2]);
+    });
+
+    afterEach(async () => {
+      await cleanTeamsTestUp(teamSetupInfo);
+    });
+
+    test('leave team test', async ()=>{
+      await testLeaveTeamUsualCase(teamSetupInfo);
+    });
+  });
+});
+
+export interface TeamInfo {
   team: Team;
   teamMembers: TeamMember[];
 }
 
-interface TeamTestSetupInfo {
+export interface TeamTestSetupInfo {
   event: Event;
   teams: TeamInfo[]; 
 }
 
-async function setTeamsTestUp(teamMemberCounts: number[]) {
-  try {
-    const teamsInfo = await prisma.$transaction(async (tx) => {  
-      if(!teamMemberCounts) {
-        console.error('[ERROR] Teams test setup cannot have empty teamMemberCounts');
-        throw new Error('teamMemberCounts cannot be empty');
-      }
-
-      const testEvent = await tx.event.create({
-        data: {
-          name: 'EVENT FOR UNIT TESTING'
-        }
-      });
-      
-      const teamsInfo: TeamInfo[] = [];
-      for(let i = 0; i < teamMemberCounts.length; i++) {
-        const memberCount = teamMemberCounts[i];
-        if (typeof memberCount !== 'number' || memberCount === null || isNaN(memberCount) || memberCount < 0) {
-          throw new Error(`Invalid team member count at index ${i}: ${memberCount}`);
-        }
-        const tempTeam = await createTeam(tx, testEvent.id);
-        const teamMembers: TeamMember[] = [];
-
-        for(let j = 0; j < memberCount; j++) {
-          const tempProfile = await tx.userProfile.create({
-            data: {
-              id: crypto.randomUUID(), //db currently uses auth.uid but we are not creating users via normal auth flow for testing
-              firstName: `TEAM ${i + 1} TEST MEMBER ${j + 1} FN`,
-              lastName: `TEAM ${i + 1} TEST MEMBER ${j + 1} LN`
-            }
-          });
-
-          const tempTeamMember = await tx.teamMember.create({
-            data: {
-              teamId: tempTeam.id,
-              eventId: testEvent.id,
-              userId: tempProfile.id,
-              isAdmin: j === 0 //if first member of team admin true otherwise false
-            }
-          });
-          teamMembers.push(tempTeamMember);
-        }
-        const teamInfo: TeamInfo = {
-          team: tempTeam,
-          teamMembers: teamMembers
-        };
-        teamsInfo.push(teamInfo);
-      }
-
-      const teamSetUpInfo: TeamTestSetupInfo = {
-        event: testEvent,
-        teams: teamsInfo
-      };
-
-      return teamSetUpInfo;
-    });
-
-    return teamsInfo;
-
-  } catch(error) {
-    console.error('fatal error setting up teams tests', error);
-    throw error;
-  } 
-}
-
-
-describe('Team tests for 2 teams. initial team state: one member per team', ()=>{
-  let teamSetupInfo: TeamTestSetupInfo;
-  beforeEach(async () => {
-    teamSetupInfo = await setTeamsTestUp([1, 1]);
-  });
-
-  afterEach(async () => {
-    await cleanTeamsTestUp(teamSetupInfo);
-  });
-
-  test('join team test', async ()=>{
-    await testJoinTeam(teamSetupInfo);
-  });
-});
-
-// Prisma does not support nested transactions so unless the functions we test start to take a transaction, we cannot simply rollback to reverse setup.
-async function cleanTeamsTestUp(setUpInfo: TeamTestSetupInfo) {
-  //IN THEORY DELETING THE EVENT SHOULD DELETE THE RELATED TEST TEAMS AND TEAM MEMEBERS VIA DATABASE CASCADES BUT PROFILES WILL REMAIN, WE CLEAN THOSE MANUALLY
-  try{
-    await prisma.$transaction(async (tx)=>{
-      await tx.event.delete({
-        where: {
-          id: setUpInfo.event.id
-        }
-      });
-
-      const profileIds = [];
-      for(const team of setUpInfo.teams) {
-        for(const member of team.teamMembers) {
-          profileIds.push(member.userId);
-        }
-      }
-      await tx.userProfile.deleteMany({
-        where: {
-          id: {
-            in: profileIds
-          }
-        }
-      });
-    });
-  } catch (error) {
-    console.error('fatal error cleaning up teams tests', error);
-    throw error;
-  } 
-}
-
-async function testJoinTeam(testInfo: TeamTestSetupInfo) {
+async function testJoinTeamOneMemPerTeam(testInfo: TeamTestSetupInfo) {
   
   const teamToJoin = testInfo.teams?.[0];
   expect(teamToJoin).toBeDefined();
@@ -170,4 +92,36 @@ async function testJoinTeam(testInfo: TeamTestSetupInfo) {
   });
   
   expect(oldTeamDb, 'old team should be deleted because there was only one member who joined a new team.').toBeNull();
+}
+
+async function testLeaveTeamUsualCase(testInfo: TeamTestSetupInfo) {
+  // one team with 2 members. Non admin gets kicked.
+  // expected passing criteria:
+  // new team shall be created with kicked member apart of it
+  // kicked member should be admin of new team
+  // kicked member shall no longer be on old team
+
+  expect(testInfo.teams, 'teams list cannot be empty').not.toBeNull();
+  const origTeam = testInfo.teams[0];
+  expect(origTeam?.teamMembers).not.toBeNull();
+  expect(origTeam?.teamMembers.length).greaterThan(1);
+
+  await kickTeamMember(origTeam?.teamMembers[0]?.userId!, origTeam?.teamMembers[1]?.userId!, origTeam?.team.id!);
+
+  const origTeamMembersDb = await prisma.teamMember.findMany({
+    where: {
+      teamId: origTeam!.team.id
+    }
+  });
+
+  expect(origTeamMembersDb, 'kicking member should not result in no team members existing on original team').not.toBeNull();
+  expect(origTeamMembersDb[0]?.isAdmin, 'remaining member should be admin').toBe(true);
+
+  const kickedTeamMemberDb = await prisma.teamMember.findFirst({
+    where: {
+      userId: origTeam?.teamMembers[1]?.userId! 
+    }
+  });
+
+  expect(kickedTeamMemberDb?.isAdmin, 'kicked team member should now be admin').toBe(true); 
 }
